31e

+++ app/src/main/java/com/example/vaccinationlistactivity/Vaccination.kt
@@ -7,7 +7,7 @@ import java.util.*
 @Parcelize
 data class Vaccination(
     val country: String,
-    val timeline: SortedMap<String, Int>
+    val timeline: SortedMap<String, Long>
     ): Parcelable, Comparable<Vaccination> {
     override fun compareTo(other: Vaccination): Int {
         TODO("Not yet implemented")
